feat: using build exit code in pipeline

+++ go.sh
@@ -1,3 +1,5 @@
 docker compose build
+if [ "$?" != "0" ]; then exit 1; fi
+
 docker compose down
 docker compose up -d
